Make sure the kernel doesn't get polluted by stray Tempesta SKBs
SKBs that are in Tempesta's possession (those are part of a Tempesta
message) may get cloned be the kernel on egress path. A cloned SKB
has its SKB header mostly inherited from the parent, and that includes
the skb->cb[] part that contains Tempesta linked list pointers among
other things. Those are the pointers that don't let the kernel delete
SKBs that are still in Tempesta's possesssion (see a check for that
in  __kfree_skb()). Tempesta doesn't know anything about those cloned
SKBs, and the kernel can't delete them now and ever due to that check.
A kernel running Tempesta gets polluted by these stray SKBs that can't
be deleted.

This addresses the issue natsys/tempesta#96
The patch clears the area in skb->cb[] that keeps these pointers.
The kernel is able to delete these SKBs now as it sees fit.